move events to GameStateEvent.cs
