+++ Assets/Nova/Sources/Core/GameStateEvents.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1c9ea1f2fb4cff84698aaabd93e0de1c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
